refactor style; code tweaks;

+++ src/apu.h
@@ -6,21 +6,22 @@ struct Memory; // Forward declaration
 
 struct Channel
 {
-    int      period_timer   = 0; // must stay int — NoiseChannel stores values up to ~917504
+    int period_timer = 0; // must stay int — NoiseChannel stores values up to ~917504
     uint16_t length_counter = 0; // max 256 (CH3), safe
-    uint8_t  length_value   = 0;
-    bool     dac_enabled    = false;
-    bool     enabled        = false;
-    bool     length_enabled = false; // Stores raw length register value for triggering
+    uint8_t length_value = 0;
+    bool dac_enabled = false;
+    bool enabled = false;
+    bool length_enabled = false; // Stores raw length register value for triggering
     void clock_length();
 };
 
-struct Envelope {
+struct Envelope
+{
     uint8_t initial_vol = 0;
     uint8_t current_vol = 0;
-    uint8_t env_pace    = 0;
+    uint8_t env_pace = 0;
     uint8_t env_counter = 0;
-    bool    env_add     = false;
+    bool env_add = false;
 };
 
 struct SquareChannel : public Channel
@@ -35,39 +36,39 @@ struct SquareChannel : public Channel
 
 struct SquareChannelWithSweep : public SquareChannel
 {
-    uint16_t shadow_period  = 0;
-    uint8_t  sweep_pace     = 0;
-    uint8_t  sweep_step     = 0;
-    uint8_t  sweep_counter  = 0;
-    bool     sweep_negate   = false;
-    bool     sweep_enabled  = false; // Track if sweep has fired
+    uint16_t shadow_period = 0;
+    uint8_t sweep_pace = 0;
+    uint8_t sweep_step = 0;
+    uint8_t sweep_counter = 0;
+    bool sweep_negate = false;
+    bool sweep_enabled = false; // Track if sweep has fired
 };
 
 struct WaveChannel : public Channel
 {
     uint16_t period_value = 0;
-    uint8_t  volume = 0; // 0=off, 1=100%, 2=50%, 3=25%
-    uint8_t  wave_pos = 0;
+    uint8_t volume = 0; // 0=off, 1=100%, 2=50%, 3=25%
+    uint8_t wave_pos = 0;
 };
 
 struct NoiseChannel : public Channel
-{   
-    uint16_t lfsr        = 0;  // 15-bit LFSR
-    uint8_t  clock_shift = 0;
-    uint8_t  clock_div   = 0;
-    uint8_t  lfsr_width  = 15; // 15-bit or 7-bit
+{
+    uint16_t lfsr = 0; // 15-bit LFSR
+    uint8_t clock_shift = 0;
+    uint8_t clock_div = 0;
+    uint8_t lfsr_width = 15; // 15-bit or 7-bit
     Envelope envelope;
 };
 
-struct APU 
+struct APU
 {
     uint16_t cycle_counter = 0;
-    uint8_t  frame_seq_counter = 0;
-    float    sample_counter = 0.0f;
-    bool     master_enabled = false;
-    uint8_t  nr50 = 0; // Master volume & VIN panning
-    uint8_t  nr51 = 0; // Sound panning
-    uint8_t  nr52 = 0; // Audio Master control
+    uint8_t frame_seq_counter = 0;
+    uint32_t sample_counter = 0; 
+    bool master_enabled = false;
+    uint8_t nr50 = 0; // Master volume & VIN panning
+    uint8_t nr51 = 0; // Sound panning
+    uint8_t nr52 = 0; // Audio Master control
 
     SquareChannelWithSweep channel1; // Square wave with sweep
     SquareChannel channel2; // Square wave without sweep
